fix(shifts/workload): address Claude bot review on PR #619
- Re-route /Admin/Workload → /Shifts/Admin/Workload per no-admin-url-section
  HARD RULE. Rename AdminWorkloadController to ShiftWorkloadAdminController
  and move views to Views/ShiftWorkloadAdmin/. Update AdminNavTree entry,
  breadcrumb, and IWorkloadService doc comment to match.

- Move display ordering from WorkloadService to ShiftWorkloadAdminController
  per display-sort-in-controllers. Service returns unsorted lists; controller
  assembles ordered view model. Drop the order-asserting service test in
  favour of value-based per-user totals (sort is no longer a service concern).

- Add docs/features/shifts/workload-dashboard.md feature spec per
  feature-spec-on-new-feature: business context, three user stories,
  hour math, routes/auth, architecture notes, deferred-scope record.

Refs nobodies-collective/Humans#734
