+++ src/Humans.Web/Views/ShiftWorkloadAdmin/Index.cshtml
@@ -8,7 +8,7 @@
 <div class="container py-4">
     <nav aria-label="breadcrumb">
         <ol class="breadcrumb">
-            <li class="breadcrumb-item"><a asp-controller="Admin" asp-action="Index">Admin</a></li>
+            <li class="breadcrumb-item"><a asp-controller="Shifts" asp-action="Index">Shifts</a></li>
             <li class="breadcrumb-item active" aria-current="page">Workload</li>
         </ol>
     </nav>
